--- a/src/common/domain.helper.ts
+++ b/src/common/domain.helper.ts
@@ -1,0 +1,16 @@
+/**
+ * Based on the data, return an array with unique values.
+ *
+ * @export
+ * @returns array
+ * @param results
+ */
+export function getUniqueXDomainValues(results: any[]): any[] {
+  const valueSet = new Set();
+  for (const result of results) {
+    for (const d of result.series) {
+      valueSet.add(d.name);
+    }
+  }
+  return Array.from(valueSet);
+}