Another one coming in. :)

Converted to using sets instead of includes on possible large data.
Tackles #714 and maybe #519, but not completely.

For the usecase (2 lines, 30k datapoints) duration went from 600ms to 80ms, the function gets called 4 times, so it makes a difference, but it still takes a bit of time to load this graph.

With 100 points, the performance is about equal, below 100 datapoints, old scenario is faster.
But we are talking about overhead of going from set --> array, which in category of a 0.00x ms differences on my machine.

It just makes ngx-charts a bit more scalable/useable with larger data (>100 datapoints).

**What kind of change does this PR introduce?** (check one with "x")
- [ ] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, local variables)
- [x] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] CI related changes
- [ ] Other... Please describe:

**What is the current behavior?** (You can also link to an open issue here)
Using Array.includes() to check for unique values.
For the usecase (2 lines, 30k datapoints): 600ms

**What is the new behavior?**
Using a Set to make sure everything is unique, and then converting to an array.
For the usecase (2 lines, 30k datapoints): 80ms

**Does this PR introduce a breaking change?** (check one with "x")
- [ ] Yes
- [x] No

If this PR contains a breaking change, please describe the impact and migration path for existing applications: ...


**Other information**: