@marjan-georgiev 2018-04-08
Awesome work! The only thing I would change is to extract this function into a service and import it from there, instead of putting it in the base class. Not all charts have an x axis, so it doesn't apply to every chart that inherits from the base chart.