@EricSmekens 2018-04-08
@marjan-georgiev Were you thinking about a service in the ChartCommonModule?
And then inject it in each component that needs to use it?
Adding to each component etc.
```
  providers: [DomainService],
```

Seperate function and importing it with a ES6 import is also an option, might be easier for as what is the functionality now, because a single service for this 1 function, I doubt about that.

But if that is the way to go, I will happily add a commit implementing 1 of those 2 options.