@marjan-georgiev 2018-04-08
I was thinking about a separate file with the function and importing it through ES6 import. Chose a confusing word there, sorry.